take logic from controller
