+++ CarService.Api/Models/AuthOptions.cs
@@ -5,8 +5,8 @@ namespace CarService.Api.Models
 {
     public class AuthOptions
     {
-        public const string ISSUER = "MyAuthServer"; // издатель токена
-        public const string AUDIENCE = "http://localhost:51884/"; // потребитель токена
+        public const string ISSUER = "CarServer"; // издатель токена
+        public const string AUDIENCE = "http://localhost:5000/"; // потребитель токена
         const string KEY = "mysupersecret_secretkey!123";   // ключ для шифрации
         public const int LIFETIME = 1; // время жизни токена - 1 минута
         public static SymmetricSecurityKey GetSymmetricSecurityKey()
