+++ CarService.Api/Models/UserDTO.cs
@@ -1,0 +1,9 @@
+namespace CarService.Api.Models
+{
+    public class UserDTO
+    {
+        
+        public string username { get; set; }
+        public string password { get; set; }
+    }
+}
