+++ CarService.Api/Services/AccountService.cs
@@ -1,7 +1,58 @@
+using System;
+using System.Security.Claims;
+using Microsoft.IdentityModel.Tokens;
+using System.IdentityModel.Tokens.Jwt;
+using CarService.Api.Models;
+using System.Collections.Generic;
+using System.Linq;
+
+
 namespace CarService.Api.Services
 {
-    public class AccountService
+
+    public class AccountService : IAccountService
     {
-        
+        // later to get rid of
+          private List<User> users = new List<User>
+        {
+            new User {Email="mechanicn@gmail.com", Password="12345", Role = "mechanic" },
+            new User { Email="qwerty", Password="55555", Role = "user" }
+        };
+
+        public string createJwtToken(ClaimsIdentity identity)
+        {
+            var now = DateTime.UtcNow;
+           
+            var jwt = new JwtSecurityToken(
+                issuer: AuthOptions.ISSUER,
+                audience: AuthOptions.AUDIENCE,
+                notBefore: now,
+                claims: identity.Claims,
+                expires: now.Add(TimeSpan.FromMinutes(AuthOptions.LIFETIME)),
+                signingCredentials: new SigningCredentials(AuthOptions.GetSymmetricSecurityKey(), SecurityAlgorithms.HmacSha256));
+            string encodedJwt = new JwtSecurityTokenHandler().WriteToken(jwt);
+            return encodedJwt;
+        }
+
+        public ClaimsIdentity GetIdentity(string username, string password)
+        {
+             User user = users.FirstOrDefault(x => x.Email == username && x.Password == password);
+            if (user != null)
+            {
+                var claims = new List<Claim>
+                {
+                    new Claim(ClaimsIdentity.DefaultNameClaimType, user.Email),
+                    new Claim(ClaimsIdentity.DefaultRoleClaimType, user.Role)
+                };
+                ClaimsIdentity claimsIdentity =
+                new ClaimsIdentity(claims, "Token", ClaimsIdentity.DefaultNameClaimType,
+                    ClaimsIdentity.DefaultRoleClaimType);
+                return claimsIdentity;
+            }
+ 
+            // если пользователя не найдено
+            return null;
+        }
+
     }
 }
